style: Fix all clippy+rustfmt lint warnings
